requirements-simple.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib
python-multipart
flet
python-dotenv
bcrypt
